Overload
changed  some dicts to enum.
tried to get better naming sense for dicts and enums.
added init function for one dictionary
changed enums back because enums have to be converted to int explicitly and that is ass.
added layerID param to UpdateTilemapLayers (renamed to be less of a mouthful)
and removed the z loop inside since it is no longer neccessary
removed unneccessary comments

changed init to overloading UpdateTileMapLayers
added a bunch of other questionable useful overload functions